requirements.txt
deepgram-sdk==v3.7.3
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
requests
langchain
langchain-community
langchain-core
langchain-openai
numpy==1.26.3
openai
pandas==2.1.4
posthog==3.3.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
langchain_groq